security: Implement certificate chains in NaiveCertificateProvider
Chain certificates are pre-populated into the cache, as they would have to be requested first otherwise, which is not even implemented, yet.

It's important to set the time when creating the NaiveCertificateProvider, as the certificate times are off otherwise. This doesn't matter for authorization tickets, as they're automatically renewed, but it applies to the root certficate and the authorization authority. It didn't apply before, because the time on trusted certificates isn't checked, yet.