security: Fix header serialization order
Sorting is done in the header policy instead of in the serialization function to (1) keep the const in serialize() and (2) keep the order flexible for both testing other implementations and further certificate profiles.

The order is specified to be always ascending, unless overridden by a certificate profile. All currently defined certificate profiles specify that signer_info must be encoded first.

Change-Id: I691a9dbe8ee26c8e0d2b641a620da4555f5d11b1