security: Check and expose permissions
The indicated ITS-AID in the message is now required to be present in the certificate. Certificates are checked for consistency of permissions with the signer. As CA certificates only contain AIDs and no SSPs, the security entity and services don't need to know about SSP formats. It's the task of the facilities layer to check payloads for compliance with the indicated AID and SSPs, which are exposed now.

The stricter checks required corresponding changes in the NaiveCertificateManager, which is extensively used in tests.