security: Optimize and improve certificate validation
Make use of the fact that certificates must be consistent. This means
this consistency must only be verified once for every certificate. As
all certificates in the certificate cache have already been verified,
these don't have to be checked again. Validation of certificates from
the cache can be completely avoided.

For this to work, certificate region and time matching has been moved
into the VerifyService.

Overall, this greatly reduces the overhead on verification of
certificates. Further optimizations could be done by also fetching
certificates in the signer_info from the cache and if they're in the
cache, skip the signature validation there, too.

Note for future a revocation implementation: If revocation should be
checked also for non-new certificates, it's probably best to extend the
certificate cache to record the chain with the certificate, otherwise a
signature validation has to happen again, because HashedId8 identifiers
aren't unique and aren't enough to check equality.

A few tests have been dropped as they no longer apply in that form.
These will be re-added in a future commit in a modified form.