socktap: update to Cohda LLC network interface
Cohda has discontinued the WAVE network interfaces on the MK5 devices
and added 'Cohda LLC' network interfaces. While the WAVE interfaces
operated with Ethernet frames, the Cohda LLC interfaces operate with
802.11p frames, so socktap has to prepend GN packets with a link layer
header (Logical Link Control) and an 802.11p header (most of the fields
of both remain constant) as well as the header for the network
interface message. Sequence control and calculation of the
frame check sequence is done by the network interface. On the receiving
side, socktap has to check for valid interface messages and valid
802.11p and link layer headers before passing the packet to the GN router.