socktap: change security_entity parameter of RouterContext to pointer
The class Router's setter for its security_entity member needs a
pointer to a SecurityEntity which may as well be nullptr. With this in
mind, RouterContext should also only deal with a pointer. This way,
if no security is wanted/needed, one does not have to create a dummy
SecurityEntity just to satisfy RouterContext's interface.